// assestmanagement/src/app/interfaces/AssestSummary.ts
export interface AssetSummary {
    assignedFixedAssets: number;
    unassignedFixedAssets: number;
    notWorkingFixedAssets: number;
    assignedItAssets: number;
    unassignedItAssets: number;
    notWorkingItAssets: number;
    totalAssets: number;
    totalAssignedAssets: number;
    totalUnassignedAssets: number;
    totalNotWorkingAssets: number;
  }
  